+++ src/main/java/com/atlauncher/constants/Constants.java
@@ -102,7 +102,9 @@ public class Constants {
     // if you fork or modify this launcher, you must not use this API key and apply
     // for your own
     public static final String CURSEFORGE_CORE_API_KEY = "$2a$10$.7CSxLm/lnj5lCBSM5jGQ.3SICSX4j9r661AgoB1Rc4Nw8jCMKcv2";
+    public static final String CURSEFORGE_API_KEY_HEADER = "x-api-key";
     public static final String CURSEFORGE_CORE_API_HOST = "api.curseforge.com";
+    public static final String CURSEFORGE_FILE_DOWNLOAD_HOST = "edge.forgecdn.net";
     public static final int CURSEFORGE_FORGE_MODLOADER_ID = 1;
     public static final int CURSEFORGE_FABRIC_MODLOADER_ID = 4;
     public static final int CURSEFORGE_QUILT_MODLOADER_ID = 5;
@@ -124,6 +126,8 @@ public class Constants {
     // Modrinth domains, endpoints, config, etc
     public static final String MODRINTH_API_URL = "https://api.modrinth.com/v2";
     public static final String MODRINTH_HOST = "api.modrinth.com";
+    public static final String MODRINTH_FILE_DOWNLOAD_HOST = "cdn.modrinth.com";
+    public static final String MODRINTH_DOWNLOAD_METADATA_HEADER = "modrinth-download-meta";
     public static final String MODRINTH_FABRIC_MOD_ID = "P7dR8mSH";
     public static final String MODRINTH_LEGACY_FABRIC_MOD_ID = "9CJED7xi";
     public static final String MODRINTH_QSL_MOD_ID = "qvIfYCYJ";